Finish getting OS X build running again
- Remove some notification code not removed when the Growl code was removed.
- Fix a uint64_t/size_t function prototype issue.

The Mac build still doesn’t quite run but this seems to be due to deeper issues. This PR just gets the splash screen to (sort of) properly load again.

+++ cppForSwig/ScriptRecipient.h
@@ -95,7 +95,7 @@ public:
    }
 
    //return size is static
-   uint64_t getSize(void) const { return 34; }
+   size_t getSize(void) const { return 34; }
 };
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -124,7 +124,7 @@ public:
       script_ = move(bw.getData());
    }
 
-   uint64_t getSize(void) const { return 31; }
+   size_t getSize(void) const { return 31; }
 };
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -154,7 +154,7 @@ public:
       script_ = move(bw.getData());
    }
 
-   uint64_t getSize(void) const { return 32; }
+   size_t getSize(void) const { return 32; }
 };
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -183,7 +183,7 @@ public:
       script_ = move(bw.getData());
    }
 
-   uint64_t getSize(void) const { return 43; }
+   size_t getSize(void) const { return 43; }
 };
 
 #endif
